feature: rename style components files and chnage location of some files
