+++ package.json
@@ -19,10 +19,10 @@
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-  "start": "vite",
-  "build": "tsc && vite build",
-  "serve": "vite preview"
-},
+    "start": "vite",
+    "build": "tsc && vite build",
+    "serve": "vite preview"
+  },
   "eslintConfig": {
     "extends": [
       "react-app",
@@ -42,6 +42,7 @@
     ]
   },
   "devDependencies": {
+    "@types/styled-components": "^5.1.26",
     "@vitejs/plugin-react": "^4.0.0",
     "autoprefixer": "^10.4.11",
     "babel-plugin-macros": "^3.1.0",
