+++ src/App.tsx
@@ -1,11 +1,11 @@
-import GlobalStyle from "./styles/generic/GlobalStyle";
+// import GlobalStyle from "./styles/generic/GlobalStyle";
 import LandingPage from "./pages/LandingPage";
 
 
 function App() {
   return (
     <div className="App">
-      <GlobalStyle />
+      {/* <GlobalStyle /> */}
        <LandingPage />
     </div>
   );
