--- src/pages/data/navbarItems.json
@@ -1,18 +1,0 @@
-[
-	{
-		"title": "about",
-		"href": "#about"
-	},
-	{
-		"title": "skillset",
-		"href": "#skillset"
-	},
-	{
-		"title": "projects",
-		"href": "#projects"
-	},
-	{
-		"title": "contact",
-		"href": "#contact"
-	}
-]
