--- src/pages/data/projects.json
@@ -1,26 +1,0 @@
-[
-	{
-		"src": "assets/images/drawingandSound.png",
-		"href": "https://drawing-sound.herokuapp.com/"
-	},
-	{
-		"src": "assets/images/inkspace.png",
-		"href": "https://inkspace.herokuapp.com/"
-	},
-	{
-		"src": "assets/images/pickitupmate.png",
-		"href": "https://rocky-gorge-25750.herokuapp.com/"
-	},
-	{
-		"src": "assets/images/asteroids1979.png",
-		"href": "https://asteroidsjavascript-imjarjjxzw.now.sh/"
-	},
-	{
-		"src": "assets/images/earthmoonsystem.png",
-		"href": "https://misaelmp.github.io/Earth_Moon_System/"
-	},
-	{
-		"src": "assets/images/portfolio2.png",
-		"href": "http://second-cover.surge.sh/"
-	}
-]
