--- src/pages/data/skillset.json
@@ -1,44 +1,0 @@
-[
-	{
-		"icon": "html5"
-	},
-	{
-		"icon": "tailwindcss"
-	},
-	{
-		"icon": "webpack"
-	},
-	{
-		"icon": "gulp"
-	},
-	{
-		"icon": "github"
-	},
-	{
-		"icon": "rails"
-	},
-	{
-		"icon": "ruby"
-	},
-	{
-		"icon": "nodejs"
-	},
-	{
-		"icon": "vuejs"
-	},
-	{
-		"icon": "javascript"
-	},
-	{
-		"icon": "react"
-	},
-	{
-		"icon": "css3"
-	},
-	{
-		"icon": "typescript"
-	},
-	{
-		"icon": "processing"
-	}
-]
