+++ src/styles/elements/navbar.tsx
