+++ src/styles/objects/section.tsx
